backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.104.1
google-adk>=0.0.1
google-genai>=1.0.0
google-generativeai>=0.8.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
openai==1.3.0
packaging==25.0
pluggy==1.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.2
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1